examples/drf_project/requirements.txt
altgraph==0.17.2
aniso8601==7.0.0
appdirs==1.4.3
apturl==0.5.2
attrs==19.3.0
bcrypt==3.1.7
black==19.3b0
blinker==1.4
Brlapi==0.7.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.9
Click==7.0
colorama==0.4.3
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.0
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==2.2.24
django-cors-headers==3.10.1
django-filter==21.1
djangorestframework==3.13.1
drf-yasg==1.20.0
duplicity==0.8.12.0
entrypoints==0.3
Faker==12.0.1
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
httplib2==0.14.0
idna==3.3
importlib-metadata==4.11.3
inflection==0.5.1
isort==4.3.4
itypes==1.2.0
Jinja2==3.1.2
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.6
MarkupSafe==2.1.1
mixer==7.2.2
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.4
oauthlib==3.1.1
packaging==21.3
paramiko==2.6.0
pbr==5.9.0
pexpect==4.6.0
Pillow==8.4.0
promise==2.3
protobuf==3.6.1
pycairo==1.16.2
pycparser==2.21
pycups==1.9.73
PyGObject==3.36.0
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.4
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.0.9
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python3-openid==3.2.0
pytz==2021.3
pyxdg==0.26
PyYAML==5.3.1
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
Rx==1.6.1
SecretStorage==2.3.1
simplejson==3.16.0
singledispatch==3.7.0
six==1.14.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
speedtest-cli==2.1.3
sqlparse==0.4.2
systemd-python==234
testresources==2.0.1
text-unidecode==1.3
toml==0.10.0
ubuntu-advantage-tools==27.8
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.7
usb-creator==0.3.7
virtualenv==20.0.17
wadllib==1.3.3
xkit==0.0.0
zipp==1.0.0
psycopg2==2.8.6
